requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.2
python-docx>=1.1.0
typing>=3.7.4.3 